requirements.txt
streamlit
openai
google-genai
firebase-admin
